gets data from local csv
