+++ public/data.csv
@@ -1,0 +1,101 @@
+date,country
+2022-6-14,cz
+2022-6-15,lk
+2022-6-16,eg
+2022-6-17,ch
+2022-6-18,vn
+2022-6-19,sd
+2022-6-20,be
+2022-6-21,kz
+2022-6-22,ye
+2022-6-23,ga
+2022-6-24,tf
+2022-6-25,ax
+2022-6-26,np
+2022-6-27,ad
+2022-6-28,bz
+2022-6-29,bt
+2022-6-30,hm
+2022-7-1,vg
+2022-7-2,fr
+2022-7-3,tn
+2022-7-4,ba
+2022-7-5,mx
+2022-7-6,kg
+2022-7-7,wf
+2022-7-8,tt
+2022-7-9,nu
+2022-7-10,cn
+2022-7-11,fo
+2022-7-12,tl
+2022-7-13,ai
+2022-7-14,gl
+2022-7-15,dj
+2022-7-16,am
+2022-7-17,bd
+2022-7-18,il
+2022-7-19,mg
+2022-7-20,si
+2022-7-21,vi
+2022-7-22,nl
+2022-7-23,by
+2022-7-24,gy
+2022-7-25,dm
+2022-7-26,sx
+2022-7-27,pw
+2022-7-28,mv
+2022-7-29,yt
+2022-7-30,fi
+2022-7-31,gh
+2022-8-1,cl
+2022-8-2,tg
+2022-8-3,bv
+2022-8-4,gb
+2022-8-5,cd
+2022-8-6,kh
+2022-8-7,mw
+2022-8-8,ae
+2022-8-9,vu
+2022-8-10,ss
+2022-8-11,ky
+2022-8-12,lv
+2022-8-13,tz
+2022-8-14,tw
+2022-8-15,cg
+2022-8-16,gq
+2022-8-17,sj
+2022-8-18,kr
+2022-8-19,ph
+2022-8-20,nc
+2022-8-21,sv
+2022-8-22,br
+2022-8-23,bh
+2022-8-24,sl
+2022-8-25,pr
+2022-8-26,ms
+2022-8-27,ua
+2022-8-28,ly
+2022-8-29,so
+2022-8-30,ws
+2022-8-31,do
+2022-9-1,hr
+2022-9-2,hn
+2022-9-3,cm
+2022-9-4,sz
+2022-9-5,mc
+2022-9-6,ml
+2022-9-7,ca
+2022-9-8,sb
+2022-9-9,qa
+2022-9-10,sn
+2022-9-11,kn
+2022-9-12,ao
+2022-9-13,na
+2022-9-14,ir
+2022-9-15,et
+2022-9-16,fk
+2022-9-17,at
+2022-9-18,bl
+2022-9-19,rw
+2022-9-20,nr
+2022-9-21,ug
